+++ com.microej.example.iot.androidconnectivity/src/main/java/com/microej/example/iot/androidconnectivity/Main.java
@@ -1,9 +1,9 @@
 /*
  * Java
  *
- * Copyright 2018 IS2T. All rights reserved.
- * For demonstration purpose only.
- * IS2T PROPRIETARY. Use is subject to license terms.
+ * Copyright 2018-2019 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ * MicroEJ Corp. PROPRIETARY. Use is subject to license terms.
  */
 package com.microej.example.iot.androidconnectivity;
 
@@ -12,14 +12,14 @@ package com.microej.example.iot.androidconnectivity;
  */
 public class Main {
 
+
 	/**
 	 * Entry point.
 	 *
 	 * @param args
 	 */
 	public static void main(String[] args) {
-		MyConnectivityExample connectivityExample = new MyConnectivityExample();
-		connectivityExample.registerConnectivityManager();
+		new ExampleBackgroundService().onStart();
 	}
 
 }