+++ com.microej.example.iot.json/src/main/java/com/microej/example/iot/json/Main.java
@@ -1,0 +1,31 @@
+/*
+ * Java
+ *
+ * Copyright 2016-2019 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ * MicroEJ Corp. PROPRIETARY. Use is subject to license terms.
+ */
+package com.microej.example.iot.json;
+
+/**
+ * This example uses the org.json.me parser provided by json.org to parse and
+ * browse a JSON file.
+ *
+ * The contents of the JSON file is taken from an example available from http://www.json.org/example.html
+ *
+ * The example then iterates over all the menuitem elements available in the popup menu and print their contents
+ *
+ */
+public class Main {
+
+	/**
+	 * Starts the background service.
+	 * 
+	 * @param args
+	 *            not used.
+	 */
+	public static void main(String[] args) {
+		new JsonBackgroundService().onStart();
+	}
+
+}