+++ com.microej.example.iot.zwave/src/main/java/com/microej/example/iot/zwave/SwitchWithReturnStateManager.java
@@ -1,10 +1,13 @@
 /*
  * Java
  *
- * Copyright 2016 IS2T. All rights reserved.
- * Use of this source code is subject to license terms.
+ * Copyright 2016-2019 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ * MicroEJ Corp. PROPRIETARY. Use is subject to license terms.
  */
 package com.microej.example.iot.zwave;
+
+import com.microej.example.iot.zwave.app.ZWaveBG;
 
 import ej.basedriver.SwitchWithReturnState;
 import ej.basedriver.event.EventHandler;
@@ -16,30 +19,26 @@ import ej.ecom.RegistrationListener;
  * Prints when a switch event occurs.
  */
 public class SwitchWithReturnStateManager
-implements RegistrationListener<SwitchWithReturnState>, EventHandler<SwitchWithReturnState, SwitchStateEvent> {
+		implements RegistrationListener<SwitchWithReturnState>, EventHandler<SwitchWithReturnState, SwitchStateEvent> {
 
 	@Override
 	public void deviceRegistered(RegistrationEvent<SwitchWithReturnState> event) {
-		System.out.println("New SwitchWithReturnState device registered " + event.getDevice());
-
+		ZWaveBG.LOGGER.info("New SwitchWithReturnState device registered " + event.getDevice()); //$NON-NLS-1$
 	}
 
 	@Override
 	public void deviceUnregistered(RegistrationEvent<SwitchWithReturnState> event) {
-		System.out.println("SwitchWithReturnState device unregistered " + event.getDevice());
-
+		ZWaveBG.LOGGER.info("SwitchWithReturnState device unregistered " + event.getDevice()); //$NON-NLS-1$
 	}
 
 	@Override
 	public void handleEvent(SwitchStateEvent event) {
-		System.out.println(event.getDevice() + " handleEvent " + event.getState());
-
+		ZWaveBG.LOGGER.info(event.getDevice() + " handleEvent " + event.getState()); //$NON-NLS-1$
 	}
 
 	@Override
 	public void handleError(SwitchStateEvent error) {
-		System.out.println(error.getDevice() + " handleError " + error.getState());
-
+		ZWaveBG.LOGGER.info(error.getDevice() + " handleError " + error.getState()); //$NON-NLS-1$
 	}
 
 }