+++ com.microej.example.iot.zwave.simulator/CHANGELOG.md
@@ -1,12 +1,28 @@
-## 1.0.1 (2017-08-07)
-Bug Fixes:
+# Changelog
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 1.0.1 - 2017-08-07
+
+### Fixed
+
   - Change the BaseDriverZWave version.
   
-## 1.0.0 (2016-12-29)
-Features:
+## 1.0.0 - 2016-12-29
+
+### Added
+
   - Change organization.
   
-## 0.1.0 (2016-10-03)
-Features:
+## 0.1.0 - 2016-10-03
+
+### Added
+
   - Initial revision.
 
+---  
+_Markdown_   
+_Copyright 2019 MicroEJ Corp. All rights reserved._   
+_Use of this source code is governed by a BSD-style license that can be found with this software._   
+_MicroEJ Corp. PROPRIETARY. Use is subject to license terms._